변경 : Appbar에서 Mobile 컴포넌트 분리
