+++ src/component/appbar/Mobile.tsx
@@ -1,0 +1,87 @@
+import React from "react";
+import { Box, IconButton, Typography, Menu, MenuItem } from "@mui/material";
+import MenuIcon from "@mui/icons-material/Menu";
+import { Search } from "@mui/icons-material";
+
+const pages = ["type", "name", "point"];
+
+const MobileAppbar = () => {
+  const [anchorElNav, setAnchorElNav] = React.useState<null | HTMLElement>(
+    null
+  );
+
+  const handleOpenNavMenu = (event: React.MouseEvent<HTMLElement>) => {
+    setAnchorElNav(event.currentTarget);
+  };
+
+  const handleCloseNavMenu = () => {
+    setAnchorElNav(null);
+  };
+
+  return (
+    <>
+      <Box sx={{ display: "flex" }}>
+        <IconButton
+          size="large"
+          aria-label="account of current user"
+          aria-controls="menu-appbar"
+          aria-haspopup="true"
+          onClick={handleOpenNavMenu}
+          color="inherit"
+        >
+          <MenuIcon />
+        </IconButton>
+        {/* 메뉴버튼 */}
+        <Menu
+          id="menu-appbar"
+          anchorEl={anchorElNav}
+          anchorOrigin={{
+            vertical: "bottom",
+            horizontal: "left",
+          }}
+          keepMounted
+          transformOrigin={{
+            vertical: "top",
+            horizontal: "left",
+          }}
+          open={Boolean(anchorElNav)}
+          onClose={handleCloseNavMenu}
+          sx={{
+            display: "block",
+          }}
+        >
+          {pages.map((page) => (
+            <MenuItem key={page} onClick={handleCloseNavMenu}>
+              <Typography textAlign="center">{page}</Typography>
+            </MenuItem>
+          ))}
+        </Menu>
+      </Box>
+      <Box
+        sx={{
+          display: "flex",
+        }}
+      >
+        <Search sx={{ display: "flex", mr: 0.5, mt: 0.5 }} />
+        <Typography
+          variant="h5"
+          noWrap
+          component="a"
+          href="#app-bar-with-responsive-menu"
+          sx={{
+            mr: 2,
+            display: "flex",
+            fontWeight: 700,
+            letterSpacing: 2,
+            color: "inherit",
+            textDecoration: "none",
+          }}
+        >
+          OfficeFinder
+        </Typography>
+      </Box>
+    </>
+  );
+};
+
+export default MobileAppbar;
