+++ src/component/appbar/Profile.tsx
@@ -1,48 +1,28 @@
 import React from "react";
-import { useTheme } from "@mui/material/styles";
 import {
-  AppBar,
   Box,
-  Toolbar,
   IconButton,
   Typography,
   Menu,
-  Container,
   Avatar,
   Tooltip,
   MenuItem,
-  useMediaQuery,
 } from "@mui/material";
-import MenuIcon from "@mui/icons-material/Menu";
-import { Search } from "@mui/icons-material";
-import Web from "../appbar/Web";
 
-const pages = ["type", "name", "point"];
 const settings = ["Profile", "Bookmark", "Reservation", "Logout"];
 
 const Profile = () => {
-  const [anchorElNav, setAnchorElNav] = React.useState<null | HTMLElement>(
-    null
-  );
   const [anchorElUser, setAnchorElUser] = React.useState<null | HTMLElement>(
     null
   );
 
-  const handleOpenNavMenu = (event: React.MouseEvent<HTMLElement>) => {
-    setAnchorElNav(event.currentTarget);
-  };
   const handleOpenUserMenu = (event: React.MouseEvent<HTMLElement>) => {
     setAnchorElUser(event.currentTarget);
-  };
-
-  const handleCloseNavMenu = () => {
-    setAnchorElNav(null);
   };
 
   const handleCloseUserMenu = () => {
     setAnchorElUser(null);
   };
-  const theme = useTheme();
   return (
     <>
       <Box>
